refactor: 아이디 닉네임 중복 체크 로직 통합
