+++ src/store/useFormStore.ts
@@ -1,13 +1,15 @@
-import create from "zustand";
+import { create } from "zustand";
 
 interface FormState {
   name: string;
   email: string;
+  userId: string;
   password: string;
   confirmPassword: string;
   phone: string;
   birthdate: string;
   setName: (name: string) => void;
+  setUserId: (userId: string) => void;
   setEmail: (email: string) => void;
   setPassword: (password: string) => void;
   setConfirmPassword: (confirmPassword: string) => void;
@@ -17,12 +19,14 @@ interface FormState {
 
 const useFormStore = create<FormState>((set) => ({
   name: "",
+  userId: "",
   email: "",
   password: "",
   confirmPassword: "",
   phone: "",
   birthdate: "",
   setName: (name) => set({ name }),
+  setUserId: (userId) => set({ userId }),
   setEmail: (email) => set({ email }),
   setPassword: (password) => set({ password }),
   setConfirmPassword: (confirmPassword) => set({ confirmPassword }),
